LayoutTests/media/modern-media-controls/mute-button/mute-button.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>MuteButton</code> class.");

const muteButton = new MuteButton({ layoutTraits: LayoutTraits.macOS });

shouldBeEqualToString("muteButton.element.localName", "button");
shouldBeTrue('muteButton.element.classList.contains("icon")');
shouldBeTrue('muteButton.element.classList.contains("mute")');
shouldBeEqualToString("muteButton.iconName", `${Icons.Volume}`);
shouldBeFalse("muteButton.muted");

const mutedButton = new MuteButton({ layoutTraits: LayoutTraits.macOS });
mutedButton.muted = true;
shouldBeTrue("mutedButton.muted");
shouldBeEqualToString("mutedButton.iconName", `${Icons.VolumeMuted}`);

scheduler.frameDidFire = function()
{
    if ([muteButton, mutedButton].some(button => button.width == 0))
        return;

    shouldBeTrue('muteButton.element.style.webkitMaskImage.includes("macOS/volume@")');
    shouldBeTrue('mutedButton.element.style.webkitMaskImage.includes("macOS/volume-mute@")');
    finishJSTest();
};

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
